# Improvements to secret sharing following review