--- a/include/ccf/crypto/symmetric_key.h
+++ b/include/ccf/crypto/symmetric_key.h
@@ -44,7 +44,7 @@ namespace crypto
       return GCM_SIZE_TAG + IV_SIZE;
     }
 
-    void set_random_iv(EntropyPtr entropy = crypto::create_entropy())
+    void set_random_iv(EntropyPtr entropy = crypto::get_entropy())
     {
       iv = entropy->random(IV_SIZE);
     }