--- a/src/node/secret_share.h
+++ b/src/node/secret_share.h
@@ -17,7 +17,7 @@ extern "C"
   /// SSS assumes that there is a function of this prototype
   int randombytes(void* buf, size_t n)
   {
-    crypto::EntropyPtr entropy = crypto::create_entropy();
+    crypto::EntropyPtr entropy = crypto::get_entropy();
     entropy->random((unsigned char*)buf, n);
     return 0;
   }