… src/node/test/secret_share.cpp