Ensure that operations applied to the secret or to the shares are constant time, without relying on compiler optimisation, minor renaming and test improvements.

This is not particularly performance sensitive, but for the record:

Before:
```
commit c557d9c4642517fb12338a22b2742aa793800c5c (HEAD -> main)
cmake -GNinja -DCOMPILE_TARGET=virtual ..

share:
===============================================================================
   Name (baseline is *)   |   Dim   |  Total ms |  ns/op  |Baseline| Ops/second
===============================================================================
           share_10s_d1 * |      10 |     0.175 |   17515 |      - |    57091.6
            share_100s_d1 |      10 |     0.210 |   20988 |  1.198 |    47646.3
           share_1000s_d1 |      10 |     0.560 |   56005 |  3.197 |    17855.5
             share_10s_d5 |      10 |     0.682 |   68176 |  3.892 |    14667.8
            share_100s_d5 |      10 |     1.093 |  109267 |  6.238 |     9151.9
           share_1000s_d5 |      10 |     3.293 |  329292 | 18.800 |     3036.8
   share_n_recover_10s_d1 |      10 |     0.178 |   17756 |  1.014 |    56318.0
  share_n_recover_100s_d1 |      10 |     0.213 |   21251 |  1.213 |    47055.1
 share_n_recover_1000s_d1 |      10 |     0.561 |   56115 |  3.204 |    17820.5
   share_n_recover_10s_d5 |      10 |     0.884 |   88377 |  5.046 |    11315.1
  share_n_recover_100s_d5 |      10 |     1.104 |  110449 |  6.306 |     9053.9
 share_n_recover_1000s_d5 |      10 |     3.306 |  330566 | 18.873 |     3025.1
===============================================================================
```

After:

```
commit e42b1d5b5bd4ff08405d3e1756d2acca20fdce56 (HEAD -> secret_sharing_review_feedback, origin/secret_sharing_review_feedback)
cmake -GNinja -DCOMPILE_TARGET=virtual ..

share:
===============================================================================
   Name (baseline is *)   |   Dim   |  Total ms |  ns/op  |Baseline| Ops/second
===============================================================================
           share_10s_d1 * |      10 |     0.176 |   17636 |      - |    56699.6
            share_100s_d1 |      10 |     0.186 |   18606 |  1.055 |    53745.8
           share_1000s_d1 |      10 |     0.706 |   70605 |  4.003 |    14163.2
             share_10s_d5 |      10 |     0.877 |   87712 |  4.973 |    11400.9
            share_100s_d5 |      10 |     0.974 |   97373 |  5.521 |    10269.7
           share_1000s_d5 |      10 |     3.778 |  377773 | 21.420 |     2647.1
   share_n_recover_10s_d1 |      10 |     0.179 |   17867 |  1.013 |    55967.9
  share_n_recover_100s_d1 |      10 |     0.227 |   22668 |  1.285 |    44114.1
 share_n_recover_1000s_d1 |      10 |     0.708 |   70829 |  4.016 |    14118.4
   share_n_recover_10s_d5 |      10 |     0.893 |   89263 |  5.061 |    11202.8
  share_n_recover_100s_d5 |      10 |     1.175 |  117471 |  6.661 |     8512.7
 share_n_recover_1000s_d5 |      10 |     3.897 |  389687 | 22.095 |     2566.2
===============================================================================
```